Added python code and URL files
Added python code (ml_alorithms.py) that has all of the ML algorithms. Added several URL files that increase in size in accordance with the number in the name of the file. The number corresponds to the URL count by million. The file url_0.txt has ~1k urls. Still need to add the accuracy, precision, and recall to the output or calculate the TP, TN, FP, FN values for each algorithm.