Added python code
Added python code (lexical.py, classifer.py) that has all of the lexical feature extraction logic. The drivel file is the classifier.py. There are several dependencies that will need to be install on either python2/3 using pip2/3. The lexical feature information still needs to be added to the url list for processing by the ML algorithms.